theme

+++ src/main/webapp/WEB-INF/views/drink/list.jsp
@@ -375,7 +375,7 @@
                                 <tr >
                                     <td>${drink.id}</td>
                                     <td>${drink.name}</td>
-                                    <td><a href="/drink/delete/${drink.id}">delete</a> </td>
+                                    <td><a class="confirm" href="/drink/delete/${drink.id}">delete</a> </td>
 
                                 </tr>
 
@@ -401,24 +401,25 @@
 <!-- /#wrapper -->
 
 <!-- jQuery -->
-<script src="../vendor/jquery/jquery.min.js"></script>
+<script src="<c:url value="/resources/vendor/jquery/jquery.min.js"/>"></script>
 
 <!-- Bootstrap Core JavaScript -->
-<script src="../vendor/popper/popper.min.js"></script>
-<script src="../vendor/bootstrap/js/bootstrap.min.js"></script>
+<script src="<c:url value="/resources/vendor/popper/popper.min.js"/>"></script>
+<script src="<c:url value="/resources/vendor/bootstrap/js/bootstrap.min.js"/>"></script>
 
 <!-- Metis Menu Plugin JavaScript -->
-<script src="../vendor/metisMenu/metisMenu.min.js"></script>
+<script src="<c:url value="/resources/vendor/metisMenu/metisMenu.min.js"/>"></script>
 
 <!-- DataTables JavaScript -->
-<script src="../vendor/datatables/js/jquery.dataTables.min.js"></script>
-<script src="../vendor/datatables/js/dataTables.bootstrap4.min.js"></script>
-<script src="../vendor/datatables-responsive/dataTables.responsive.js"></script>
+<script src="<c:url value="/resources/vendor/datatables/js/jquery.dataTables.min.js"/>"></script>
+<script src="<c:url value="/resources/vendor/datatables/js/dataTables.bootstrap4.min.js"/>"></script>
+<script src="<c:url value="/resources/vendor/datatables-responsive/dataTables.responsive.js"/>"></script>
 
 <!-- Custom Theme JavaScript -->
-<script src="../dist/js/sb-admin-2.js"></script>
+<script src="<c:url value="/resources/dist/js/sb-admin-2.js"/>"></script>
 
 <!-- Page-Level Demo Scripts - Tables - Use for reference -->
+
 <script>
     $(document).ready(function () {
         $('#dataTables-example').DataTable({
